pyproject.toml
[tool.poetry]
name = "tap-sharepointsites"
version = "0.0.1"
description = "`tap-sharepointsites` is a Singer tap for sharepointsites, built with the Meltano Singer SDK."
authors = ["Henning Holgersen"]
keywords = [
    "ELT",
    "sharepointsites",
]
license = "Apache 2.0"

[tool.poetry.dependencies]
python = "<3.11,>=3.7.1"
requests = "^2.25.1"
singer-sdk = "^0.11.1"
azure-identity = "^1.11"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"
tox = "^3.24.4"
flake8 = "^3.9.2"
black = "^22.3.0"
pydocstyle = "^6.1.1"
mypy = "^0.910"
types-requests = "^2.26.1"
isort = "^5.10.1"

[tool.isort]
profile = "black"
multi_line_output = 3 # Vertical Hanging Indent
src_paths = "tap_sharepointsites"

[build-system]
requires = ["poetry-core>=1.0.8"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
# CLI declaration
tap-sharepointsites = 'tap_sharepointsites.tap:Tapsharepointsites.cli'
